Retain user group app assignments (#330)
Add option to retain `okta_app_group_assignment` and `okta_app_user` on destroy.

+++ okta/resource_okta_app_group_assignment.go
@@ -26,6 +26,7 @@ func resourceAppGroupAssignment() *schema.Resource {
 
 				_ = d.Set("app_id", parts[0])
 				_ = d.Set("group_id", parts[1])
+				_ = d.Set("retain_assignment", false)
 
 				assignment, _, err := getOktaClientFromMetadata(m).Application.
 					GetApplicationGroupAssignment(ctx, parts[0], parts[1], nil)
@@ -64,6 +65,12 @@ func resourceAppGroupAssignment() *schema.Resource {
 				DiffSuppressFunc: func(k, old, new string, d *schema.ResourceData) bool {
 					return new == ""
 				},
+			},
+			"retain_assignment": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+				Description: "Retain the group assignment on destroy. If set to true, the resource will be removed from state but not from the Okta app.",
 			},
 		},
 	}
@@ -121,6 +128,12 @@ func resourceAppGroupAssignmentRead(ctx context.Context, d *schema.ResourceData,
 }
 
 func resourceAppGroupAssignmentDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+	retain := d.Get("retain_assignment").(bool)
+	if retain {
+		// The assignment should be retained, bail before DeleteApplicationGroupAssignment is called
+		return nil
+	}
+
 	_, err := getOktaClientFromMetadata(m).Application.DeleteApplicationGroupAssignment(
 		ctx,
 		d.Get("app_id").(string),
